SITES-21278: Adds more extensibility examples
